+++ tests/Unit/Public/Get-IdoItCategory.tests.ps1
@@ -13,7 +13,7 @@ BeforeAll {
     . $testHelpersPath/MockData_Cmdb_object_read.ps1
     . $testHelpersPath/MockData_Cmdb_object_types_read.ps1
     . $testHelpersPath/MockData_cmdb_object_type_categories_read.ps1
-    . $testHelpersPath/MockData_cmbd_category_read.ps1
+    . $testHelpersPath/MockData_cmdb_category_read.ps1
     . $testHelpersPath/MockData_idoit_constants_read.ps1
     . $testHelpersPath/MockDefaultMockAtEnd.ps1
 }