+++ tests/Unit/Public/Get-IdoitDialog.tests.ps1
@@ -10,12 +10,12 @@ BeforeAll {
     $testRoot = Join-Path -Path (Get-SamplerAbsolutePath) -ChildPath 'tests'
     $testHelpersPath = Join-Path -Path $testRoot -ChildPath 'Unit/Helpers'
     . $testHelpersPath/MockConnectIdoIt.ps1
-    . $testHelpersPath/Mockdata_cmbd_dialog.ps1
+    . $testHelpersPath/Mockdata_cmdb_dialog.ps1
     . $testHelpersPath/MockData_Cmdb_object_read.ps1
     . $testHelpersPath/MockData_Cmdb_object_read.ps1
     . $testHelpersPath/MockData_Cmdb_object_types_read.ps1
     . $testHelpersPath/MockData_cmdb_object_type_categories_read.ps1
-    . $testHelpersPath/MockData_cmbd_category_read.ps1
+    . $testHelpersPath/MockData_cmdb_category_read.ps1
     . $testHelpersPath/MockData_idoit_constants_read.ps1
     . $testHelpersPath/MockDefaultMockAtEnd.ps1
 }