+++ tests/Unit/Public/Search-IdoItObject.tests.ps1
@@ -10,7 +10,8 @@ BeforeAll {
     $testRoot = Join-Path -Path (Get-SamplerAbsolutePath) -ChildPath 'tests'
     $testHelpersPath = Join-Path -Path $testRoot -ChildPath 'Unit\Helpers'
     . $testHelpersPath/MockConnectIdoIt.ps1
-    . $testHelpersPath/Mockdata_cmbd_condition_read.ps1
+    . $testHelpersPath/Mockdata_cmdb_condition_read.ps1
+    . $testHelpersPath/Mockdata_idoit_search_read.ps1
     . $testHelpersPath/MockDefaultMockAtEnd.ps1
 }
 
@@ -23,20 +24,35 @@ AfterAll {
 }
 
 Describe 'Search-IdoitObject' {
-    It 'Should find the object' {
-        $ret = Search-IdoItObject -Conditions @(
-            @{ "property" = "C__CATG__GLOBAL-title"; "comparison" = "like"; "value" = "*r540*" },
-            @{ "property" = "C__CATG__GLOBAL-type"; "comparison" = "="; "value" = "5" }
-        )
-        $ret | Should -Not -BeNullOrEmpty
-        $ret.id | Should -Be 540
-        $ret.title | Should -Be 'server540'
+    Context 'Search by conditions' {
+        It 'Should find the object' {
+            $ret = Search-IdoItObject -Conditions @(
+                @{ "property" = "C__CATG__GLOBAL-title"; "comparison" = "like"; "value" = "*r540*" },
+                @{ "property" = "C__CATG__GLOBAL-type"; "comparison" = "="; "value" = "5" }
+            )
+            $ret | Should -Not -BeNullOrEmpty
+            $ret.id | Should -Be 540
+            $ret.title | Should -Be 'server540'
+        }
+        It 'Should return empty list when no object found' {
+            $ret = Search-IdoItObject -Conditions @(
+                @{ "property" = "C__CATG__GLOBAL-title"; "comparison" = "like"; "value" = "*unknown*" },
+                @{ "property" = "C__CATG__GLOBAL-type"; "comparison" = "="; "value" = "5" }
+            )
+            $ret | Should -BeNullOrEmpty
+        }
     }
-    It 'Should return empty list when no object found' {
-        $ret = Search-IdoItObject -Conditions @(
-            @{ "property" = "C__CATG__GLOBAL-title"; "comparison" = "like"; "value" = "*unknown*" },
-            @{ "property" = "C__CATG__GLOBAL-type"; "comparison" = "="; "value" = "5" }
-        )
-        $ret | Should -BeNullOrEmpty
+    Context 'Search by query' {
+        It 'Should find the object' {
+            $ret = Search-IdoItObject -Query '*540*'
+            $ret | ForEach-Object {
+                $_.documentId | Should -Be 540
+                $_.value | Should -Match '540'
+            }
+        }
+        It 'Should return empty list when no object found' {
+            $ret = Search-IdoItObject -Query 'unknown'
+            $ret | Should -BeNullOrEmpty
+        }
     }
 }